all done :> changed float to double

+++ src/main/java/hu/petrik/szuperhosprojekt/Batman.java
@@ -18,7 +18,7 @@ public class Batman implements Szuperhos, Milliardos {
     }
 
     @Override
-    public float mekkoraAzEreje() {
+    public double mekkoraAzEreje() {
         return lelemenyesseg*2;
     }
