+++ src/main/java/hu/petrik/szuperhosprojekt/Szuperhos.java
@@ -2,6 +2,7 @@ package hu.petrik.szuperhosprojekt;
 
 public interface Szuperhos {
     public boolean legyoziE(Szuperhos s);
-    public float mekkoraAzEreje();
+
+    public double mekkoraAzEreje();
 
 }
